requirements.txt
argcomplete==1.12.3
asgiref==3.4.1
boto3==1.18.43
botocore==1.21.43
certifi==2021.5.30
cfn-flip==1.2.3
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
Django==3.2.7
djangorestframework==3.12.4
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.2
jmespath==0.10.0
kappa==0.6.0
pep517==0.11.0
pip-tools==6.2.0
placebo==0.9.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tqdm==4.62.2
troposphere<3.0.3
urllib3==1.26.6
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.53.0
